refactor: move waitServerReady to src/test-helpers and exclude from bundle
